backend/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
flask==3.0.2 